% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection-package.R
\docType{package}
\name{modelselection-package}
\alias{modelselection}
\alias{modelselection-package}
\title{modelselection: Model Selection and Tuning Utilities}
\description{
Provides common model selection and tuning utilities in an intuitive and extremely flexible manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.dmolitor.com/modelselection/}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
