% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltuning-package.R
\docType{package}
\name{modeltuning-package}
\alias{modeltuning}
\alias{modeltuning-package}
\title{modeltuning: Model Selection and Tuning Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a lightweight framework for model selection and hyperparameter tuning in R. The package offers intuitive tools for grid search, cross-validation, and combined grid search with cross-validation that work seamlessly with virtually any modeling package. Designed for flexibility and ease of use, it standardizes tuning workflows while remaining fully compatible with a wide range of model interfaces and estimation functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.dmolitor.com/modeltuning/}
  \item Report bugs at \url{https://github.com/dmolitor/modeltuning/issues}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

Authors:
\itemize{
  \item Daniel Molitor \email{molitdj97@gmail.com}
}

}
\keyword{internal}
