% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelselection-package.R
\docType{package}
\name{modelselection-package}
\alias{modelselection}
\alias{modelselection-package}
\title{modelselection: What the Package Does (One Line, Title Case)}
\description{
Provides model selection and tuning utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.dmolitor.com/modelselection/}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
