% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltuning-package.R
\docType{package}
\name{modeltuning-package}
\alias{modeltuning}
\alias{modeltuning-package}
\title{modeltuning: Model Selection and Tuning Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides common model selection and tuning utilities in an intuitive and extremely flexible manner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.dmolitor.com/modeltuning/}
}

}
\author{
\strong{Maintainer}: Daniel Molitor \email{molitdj97@gmail.com}

}
\keyword{internal}
